requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
pydantic[email]==2.5.2
PyJWT==2.8.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
google-auth==2.25.2
google-cloud-storage==2.14.0
requests==2.31.0
sendgrid==6.11.0
